Created schema files for each model, and created initial readme file for running instructions
