+++ api/README.md
@@ -1,0 +1,17 @@
+<h1>API</h1>
+TODO: Description
+
+<h3>Installing Dependencies</h3>
+Run pipenv installer in api directory<br>
+`pipenv install`
+<br>
+<h3>Running the API</h3>
+Set Python Path<br><br>
+Linux & MacOS<br>
+`export PYTHONPATH=/.../tracker`
+<br><br>
+Windows:
+`set PYTHONPATH=/.../tracker`
+<br><br>
+To run the application ensure you are in the tracker directory and run:<br>
+`python3 api/app.py`
