+++ api/schemas/admin_affiliations.py
@@ -1,0 +1,16 @@
+import graphene
+from graphene import relay
+from graphene_sqlalchemy import SQLAlchemyObjectType
+
+from api.models import Admin_affiliations as Admin_aff_Model
+
+
+class AdminAffiliations(SQLAlchemyObjectType):
+	class Meta:
+		model = Admin_aff_Model
+		interfaces = (relay.Node, )
+
+
+class AdminAffConnection(relay.Connection):
+	class Meta:
+		node = AdminAffiliations
